fixed all the dtype errors

+++ calib/utils.py
@@ -116,7 +116,7 @@ class ListDataset(Dataset):
             inputs[0] = self.transform(inputs[0])
             inputs[1] = self.transform(inputs[1])
 
-        return inputs, yaw, pitch
+        return inputs, np.float32(yaw), np.float32(pitch)
 
     def __len__(self):
         return len(self.path_list)
